+++ R/minimum_wage.R
@@ -1,0 +1,18 @@
+#' Example nominal US federal minimum wage data
+#'
+#' @docType data
+#' @name us_minimum_wage
+#' @keywords datasets
+#'
+#' @format The data frame `us_minimum_wage` contains the
+#' nominal value of the US federal minimum wage since 1938. It has two columns:
+#' \describe{
+#'   \item{year}{numeric year}
+#'   \item{minimum_wage}{maximum value of the US federal minimum wage that year}
+#' }
+#'
+#' @source \href{https://www.dol.gov/agencies/whd/minimum-wage/history/chart}{US Department of Labor}
+#'
+#' @examples
+#' us_minimum_wage
+"us_minimum_wage"