+++ data-raw/R/cpi_u_x1.R
@@ -1,5 +1,5 @@
 read_cpi_u_x1 <- function(csv) {
-  read_csv(csv) %>%
+  read_csv(csv, show_col_types = FALSE) %>%
     rename_all(tolower) %>%
     select(-aavg) %>%
     pivot_longer(-year, names_to = "month", values_to = "cpi_u_x1") %>%