+++ data-raw/packages.R
@@ -2,9 +2,10 @@
 library(targets)
 library(tarchetypes)
 
-# conflicts
+# conflicts and other options
 library(conflicted)
 conflict_prefer("filter", "dplyr", quiet = TRUE)
+options(usethis.quiet = TRUE)
 
 # packages for this analysis
 suppressPackageStartupMessages({