+++ NEWS.md
@@ -1,3 +1,8 @@
+# realtalk 0.2.5
+* added extended Chained CPI-U
+* updated series through August 2023 where applicable
+* added example nominal minimum wage series
+
 # realtalk 0.2.4
 * added May 2023 CPI-U and C-CPI-U series
 
@@ -8,13 +13,10 @@
 * Added Chained CPI-U (C-CPI-U)
 
 # realtalk 0.2.1
-
 * Added `get_price_index()`
 
 # realtalk 0.2.0
-
 * Actually made the package sort of useful and documented data.
 
 # realtalk 0.1.0
-
 * The beginning